SE-299 changed unit tests and its directory
