+++ sdk/src/test/java/com/finbourne/features/FeatureFileWriter.java
@@ -1,0 +1,17 @@
+package com.finbourne.features;
+
+import java.io.*;
+
+public class FeatureFileWriter {
+
+    FileOutputStream fileOut;
+
+    public void writeToFile(String data, String fileName) throws IOException {
+        fileOut = new FileOutputStream(fileName);
+        fileOut.write(data.getBytes());
+        fileOut.close();
+    }
+
+
+
+}
