--- sdk/src/test/java/com/finbourne/features/ParityFileWriter.java
@@ -1,17 +1,0 @@
-package com.finbourne.features;
-
-import java.io.*;
-
-public class ParityFileWriter {
-
-    FileOutputStream fileOut;
-
-    public void writeToFile(String data) throws IOException {
-        fileOut = new FileOutputStream("features.txt");
-        fileOut.write(data.getBytes());
-        fileOut.close();
-    }
-
-
-
-}
